+++ Tests/Unit/MSFT_DFSReplicationGroup.Tests.ps1
@@ -96,6 +96,26 @@ try
             DomainName = 'contoso.com'
         }
 
+        $replicationGroupNullMembers = [PSObject]@{
+            GroupName = 'Test Group'
+            Ensure = 'Present'
+            Description = 'Test Description'
+            Members = $Null
+            Folders = @('Folder1','Folder2')
+            Topology = 'Manual'
+            DomainName = 'contoso.com'
+        }
+
+        $replicationGroupNullFolders = [PSObject]@{
+            GroupName = 'Test Group'
+            Ensure = 'Present'
+            Description = 'Test Description'
+            Members = @('FileServer1.contoso.com','FileServer2')
+            Folders = $Null
+            Topology = 'Manual'
+            DomainName = 'contoso.com'
+        }
+
         $replicationGroupConnections = @(
             [PSObject]@{
                 GroupName = 'Test Group'
@@ -977,6 +997,40 @@ try
                 }
             }
 
+            Context 'Replication Group exists but no members passed in' {
+                Mock Get-DfsReplicationGroup -MockWith { @($mockReplicationGroup) }
+                Mock Get-DfsrMember -MockWith { return $mockReplicationGroupMember }
+                Mock Get-DfsReplicatedFolder -MockWith { return $mockReplicationGroupFolder }
+
+                It 'Should return false' {
+                    $splat = $replicationGroupNullMembers.Clone()
+                    Test-TargetResource @splat | Should -Be $False
+                }
+
+                It 'Should call expected Mocks' {
+                    Assert-MockCalled -commandName Get-DfsReplicationGroup -Exactly -Times 1
+                    Assert-MockCalled -commandName Get-DfsrMember -Exactly -Times 1
+                    Assert-MockCalled -commandName Get-DfsReplicatedFolder -Exactly -Times 1
+                }
+            }
+
+            Context 'Replication Group exists with no members' {
+                Mock Get-DfsReplicationGroup -MockWith { @($mockReplicationGroup) }
+                Mock Get-DfsrMember -MockWith { return $null }
+                Mock Get-DfsReplicatedFolder -MockWith { return $mockReplicationGroupFolder }
+
+                It 'Should return false' {
+                    $splat = $replicationGroupNullMembers.Clone()
+                    Test-TargetResource @splat | Should -Be $True
+                }
+
+                It 'Should call expected Mocks' {
+                    Assert-MockCalled -commandName Get-DfsReplicationGroup -Exactly -Times 1
+                    Assert-MockCalled -commandName Get-DfsrMember -Exactly -Times 1
+                    Assert-MockCalled -commandName Get-DfsReplicatedFolder -Exactly -Times 1
+                }
+            }
+
             Context 'Replication Group exists but some Members passed as FQDN' {
                 Mock Get-DfsReplicationGroup -MockWith { @($mockReplicationGroup) }
                 Mock Get-DfsrMember -MockWith { return $mockReplicationGroupMember }
@@ -1065,6 +1119,40 @@ try
                     Assert-MockCalled -commandName Get-DfsReplicatedFolder -Exactly -Times 1
                 }
             }
+            Context 'Replication Group exists but no folders passed in' {
+                Mock Get-DfsReplicationGroup -MockWith { @($mockReplicationGroup) }
+                Mock Get-DfsrMember -MockWith { return $mockReplicationGroupMember }
+                Mock Get-DfsReplicatedFolder -MockWith { return $mockReplicationGroupFolder }
+
+                It 'Should return false' {
+                    $splat = $replicationGroupNullFolders.Clone()
+                    Test-TargetResource @splat | Should -Be $False
+                }
+
+                It 'Should call expected Mocks' {
+                    Assert-MockCalled -commandName Get-DfsReplicationGroup -Exactly -Times 1
+                    Assert-MockCalled -commandName Get-DfsrMember -Exactly -Times 1
+                    Assert-MockCalled -commandName Get-DfsReplicatedFolder -Exactly -Times 1
+                }
+            }
+
+            Context 'Replication Group exists with no folders' {
+                Mock Get-DfsReplicationGroup -MockWith { @($mockReplicationGroup) }
+                Mock Get-DfsrMember -MockWith { return $mockReplicationGroupMember }
+                Mock Get-DfsReplicatedFolder -MockWith { return $null }
+
+                It 'Should return false' {
+                    $splat = $replicationGroupNullFolders.Clone()
+                    Test-TargetResource @splat | Should -Be $True
+                }
+
+                It 'Should call expected Mocks' {
+                    Assert-MockCalled -commandName Get-DfsReplicationGroup -Exactly -Times 1
+                    Assert-MockCalled -commandName Get-DfsrMember -Exactly -Times 1
+                    Assert-MockCalled -commandName Get-DfsReplicatedFolder -Exactly -Times 1
+                }
+            }
+
 
             Context 'Replication Group exists but should not' {
                 Mock Get-DfsReplicationGroup -MockWith { @($mockReplicationGroup) }