+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added Hub and Spoke replication group example - fixes [Issue #62](https://github.com/PowerShell/DFSDsc/issues/62).
 - Enabled PSSA rule violations to fail build - Fixes [Issue #320](https://github.com/PowerShell/DFSDsc/issues/59).
+- Resolved issue with null values in resource group members or folders - fixes [Issue #27](https://github.com/PowerShell/xDFS/issues/27).
 
 ## 4.0.0.0